[Broadcom]Upgrade xgs SAI version to 13.2.1.1 (#23281)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
Upgrade xgs SAI to 13.2.1.1 to include the SAI_13.2.1_EA Code Drop.

##### Work item tracking
- Microsoft ADO **(number only)**: 33720110

#### How I did it
Update xgs SAI version in sai.mk file.
#### How to verify it

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [x] 202412
- [x] 202505

Signed-off-by: zitingguo-ms <zitingguo@microsoft.com>