[202311][TACACS] Fix memory leak when authenticating using tacacs (#23226)
Why I did it
A small memory leak was found when running TACACS

How I did it
A patch was created for the libnss-tacplus repository that will fix the memory leak.

How to verify it
Configure TACACS
Use the python script that reproduce the consumption and monitor its memory consumption.