[nvidia-bluefield] Fix DOCA_DPDK_BRIDGE_DEV dependencies (#23495)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
This PR is to fix the build issue while installing libdoca-sdk-dpdk-bridge-dev. This package must have libdoca-sdk-common-dev to be installed. Since that is not added as a dependency in makefile in some instances build is failing.

```
01:00:02 libdoca-sdk-dpdk-bridge-dev depends on libdoca-sdk-common-dev (>= 3.0.0058-1); however:
01:00:02 Package libdoca-sdk-common-dev is not installed.
```
##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

Explicitly add the dependency in the Makefile

#### How to verify it
Build for nvidia-bluefield platforms should not have any errors

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->